Multi-Shoveler Open and Close (#131)
* start up multiple reflectors
* Concurrent ldb downloads
* Add rotating reader and tests
* add convenience methods
* use go 1.13 friendly atomic interface
* add support for multiple LDB copies in InitContainer
* only create dbs if they don't exist
* reopen and close last rotated reader
* create a changelog for first ldb in multimode

+++ scripts/download.sh
@@ -7,28 +7,33 @@ PREFIX="$(echo $CTLSTORE_BOOTSTRAP_URL | grep :// | sed -e's,^\(.*://\).*,\1,g')
 URL="$(echo $CTLSTORE_BOOTSTRAP_URL | sed -e s,$PREFIX,,g)"
 BUCKET="$(echo $URL | grep / | cut -d/ -f1)"
 KEY="$(echo $URL | grep / | cut -d/ -f2)"
+CTLSTORE_DIR="/var/spool/ctlstore"
 CONCURRENCY=${2:-20}
+NUM_LDB=${3:-1}
 DOWNLOADED="false"
 COMPRESSED="false"
-METRICS="/var/spool/ctlstore/metrics.json"
+METRICS="$CTLSTORE_DIR/metrics.json"
 
+mkdir -p $CTLSTORE_DIR
+cd $CTLSTORE_DIR
+
+# busybox does not support sub-second resolution
 START=$(date +%s)
 END=$(date +%s)
 SHA_START=$(date +%s)
 SHA_END=$(date +%s)
 
 get_head_object() {
- head_object=$(aws s3api head-object --bucket "${BUCKET}" --key "${KEY}")
- echo "$head_object"
+  head_object=$(aws s3api head-object --bucket "${BUCKET}" --key "${KEY}")
+  echo "$head_object"
 }
 
-if [ ! -f /var/spool/ctlstore/ldb.db ]; then
-  # busybox does not support sub-second resolution
-  START=$(date +%s)
+cleanup() {
+  echo "Removing snapshot.db"
+  rm -f $CTLSTORE_DIR/snapshot.*
+}
 
-  mkdir -p /var/spool/ctlstore
-  cd /var/spool/ctlstore
-
+download_snapshot() {
   echo "Downloading head object from ${CTLSTORE_BOOTSTRAP_URL}"
   head_object=$(get_head_object)
 
@@ -44,15 +49,17 @@ if [ ! -f /var/spool/ctlstore/ldb.db ]; then
   DOWNLOADED="true"
   if [[ ${CTLSTORE_BOOTSTRAP_URL: -2} == gz ]]; then
     echo "Decompressing"
-    pigz -d snapshot.db.gz
+    pigz -df snapshot.db.gz
     COMPRESSED="true"
   fi
+}
 
+check_sha() {
   SHA_START=$(date +%s)
   if [ -z $remote_checksum ]; then
     echo "Remote checksum sha1 is null, skipping checksum validation"
   else
-    local_checksum=$(shasum snapshot.db | cut -f1 -d\ | xxd -r -p | base64)
+    local_checksum=$(shasum snapshot.db | cut -f1 -d\  | xxd -r -p | base64)
     echo "Local snapshot checksum in sha1: $local_checksum"
 
     if [[ "$local_checksum" == "$remote_checksum" ]]; then
@@ -60,11 +67,27 @@ if [ ! -f /var/spool/ctlstore/ldb.db ]; then
     else
       echo "Checksum does not match"
       echo "Failed to download intact snapshot"
+      cleanup
       exit 1
     fi
   fi
   SHA_END=$(date +%s)
   echo "Local checksum calculation took $(($SHA_END - $SHA_START)) seconds"
+}
+
+if [ ! -f "$CTLSTORE_DIR/ldb.db" ]; then
+  echo "No ldb found, downloading snapshot"
+  download_snapshot
+  check_sha
+
+  i=2
+  while [ "$i" -le $NUM_LDB ]; do
+    if [ ! -f ldb-$i.db ]; then
+      echo "creating copy ldb-$i.db"
+      cp snapshot.db ldb-$i.db
+    fi
+    i=$((i + 1))
+  done
 
   mv snapshot.db ldb.db
   END=$(date +%s)
@@ -73,5 +96,27 @@ else
   echo "Snapshot already present"
 fi
 
-echo "{\"startTime\": $(($END - $START)), \"downloaded\": \"$DOWNLOADED\", \"compressed\": \"$COMPRESSED\"}" > $METRICS
+# on existing nodes, we may already have the ldb file.
+# We should download a new snapshot to avoid copying an in-use ldb.db file and risking a malformed db
+i=2
+while [ "$i" -le $NUM_LDB ]; do
+
+  # make sure it's not already downloaded
+  if [ ! -f ldb-$i.db ]; then
+    echo "Preparing ldb-$i.db"
+    # download the snapshot if it's not present
+    if [ ! -f "$CTLSTORE_DIR/snapshot.db" ]; then
+      download_snapshot
+      check_sha
+    fi
+
+    echo "creating copy ldb-$i.db"
+    cp snapshot.db ldb-$i.db
+  fi
+  i=$((i + 1))
+done
+
+cleanup
+
+echo "{\"startTime\": $(($END - $START)), \"downloaded\": \"$DOWNLOADED\", \"compressed\": \"$COMPRESSED\"}" >$METRICS
 cat $METRICS
